Added support for Geo-Locations
Now it's possible to add locations by their latitude and longitude values to the world map. At the moment the location indicator is a simple circle but I will change this in the future so that one can use icons for the indicator.
The locations are in a separate pane (locationLayer) which can be switched on/off. If you want to see the tooltips of the locations you have to setCountryInteractionEnabled(false).